+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Support `tiff` files for icon.
 ### Updated
 - Update `GetBitmapFrame` to round `width` that is changed by `dpi`.
+- Add `GetBitmapFrameByDpiAndWidth` to get optimal frame by `dpi` and `width`.
 ### Example
 - Add `Cube-Grey-Light.tiff` and `Cube-Grey-Dark.tiff` in `AppTheme`.